project/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.1
docker==7.1.0
docutils==0.21.2
doker==0.2.0
filelock==3.14.0
idna==3.7
install==1.3.5
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
pdfrw==0.4
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
reportlab==4.2.0
requests==2.32.3
six==1.16.0
smartypants==2.0.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
